requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
openpyxl
streamlit
statsmodels
plotly
pyarrow
pyyaml 
torch
scikit-learn
joblib
